requirements.txt
Django~=3.1.1
psycopg2-binary==2.8.5
djangorestframework~=3.11.0
cffi~=1.14.0
future~=0.18.2
Pillow~=7.1.2
numpy~=1.18.1
pip~=20.2.3
wheel~=0.34.2
cryptography~=2.9.2
py~=1.8.1
pytz~=2019.3
sqlparse~=0.3.1
asgiref~=3.2.10
Jinja2~=2.11.2
setuptools~=46.1.3